impl bbq mpsc
